refactor: move networks into common
