+++ src/implementations/config/__tests__/create.test.ts
@@ -3,9 +3,9 @@ import { prompt } from "inquirer";
 import tmp from "tmp";
 import { CreateConfigCommand } from "../../../commands/config/config.type";
 import { handler as createTempDNS } from "../../../commands/dns/txt-record/create";
-import { NetworkCmdName } from "../../../commands/networks";
-import { deployDocumentStore } from "../../deploy/document-store/document-store";
-import { deployTokenRegistry } from "../../deploy/token-registry/token-registry";
+import { NetworkCmdName } from "../../../common/networks";
+import { deployDocumentStore } from "../../deploy/document-store";
+import { deployTokenRegistry } from "../../deploy/token-registry";
 import { create as createConfig } from "../create";
 import expectedConfigFileOutput from "./expected-config-file-output-v2.json";
 import inputConfigFile from "./input-config-file.json";
